Remove superfluous flags from Arduino CLI commands
The `--additional-urls` flag allows the user to supply from the command line the URLs of package indexes for platforms
that are not provided via the primary package index.

During the early phase of Arduino CLI's development, I believe it was necessary to add this flag in every command that
involved the dependent platform, and this is why the flag was added to the root command. However, that onerous
requirement was removed long ago and now the flag is only required in the commands that use the URL.

In this project, the only Arduino CLI command that uses the URL is `core install`. Inclusion of the `--additional-urls`
flag in the other commands doesn't cause any problems, but it also doesn't provide any benefit.

The `--additional-urls` flag is hereby removed from the commands where it is not required. This avoids unnecessary
length and complexity in the commands.